+++ tools/exports/song-game.js
@@ -1,0 +1,273 @@
+// This music has been exported by Voxby, a tracker built on SoundBox.
+// You can use it with https://sb.bitsnbites.eu/player-small.js in your
+// own product; see https://sb.bitsnbites.eu/demo.html for an example.
+
+// Battle · 146 bpm · 8 bars · seed 2214127632
+
+// Song data
+export default {
+      songData: [
+        { // Instrument 0
+          i: [
+          0, // OSC1_WAVEFORM
+          193, // OSC1_VOL
+          116, // OSC1_SEMI
+          64, // OSC1_XENV
+          0, // OSC2_WAVEFORM
+          193, // OSC2_VOL
+          120, // OSC2_SEMI
+          0, // OSC2_DETUNE
+          64, // OSC2_XENV
+          96, // NOISE_VOL
+          4, // ENV_ATTACK
+          6, // ENV_SUSTAIN
+          35, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          0, // LFO_AMT
+          0, // LFO_FREQ
+          0, // LFO_FX_FREQ
+          2, // FX_FILTER
+          14, // FX_FREQ
+          0, // FX_RESONANCE
+          10, // FX_DIST
+          32, // FX_DRIVE
+          0, // FX_PAN_AMT
+          0, // FX_PAN_FREQ
+          0, // FX_DELAY_AMT
+          0 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,1,1,1],
+          // Columns
+          c: [
+            {n: [135,,135,,,,,,,,135,,,,,,135,,135,,,,,,,,135],
+             f: []}
+          ]
+        },
+        { // Instrument 1
+          i: [
+          3, // OSC1_WAVEFORM
+          0, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          3, // OSC2_WAVEFORM
+          57, // OSC2_VOL
+          128, // OSC2_SEMI
+          0, // OSC2_DETUNE
+          64, // OSC2_XENV
+          183, // NOISE_VOL
+          4, // ENV_ATTACK
+          4, // ENV_SUSTAIN
+          53, // ENV_RELEASE
+          68, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          1, // LFO_WAVEFORM
+          55, // LFO_AMT
+          4, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          67, // FX_FREQ
+          115, // FX_RESONANCE
+          124, // FX_DIST
+          190, // FX_DRIVE
+          67, // FX_PAN_AMT
+          6, // FX_PAN_FREQ
+          39, // FX_DELAY_AMT
+          1 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,1,1,1],
+          // Columns
+          c: [
+            {n: [,,,,147,,,147,,,,,147,,,,,,,,147,,,147,,,,,147],
+             f: []}
+          ]
+        },
+        { // Instrument 2
+          i: [
+          2, // OSC1_WAVEFORM
+          10, // OSC1_VOL
+          140, // OSC1_SEMI
+          64, // OSC1_XENV
+          0, // OSC2_WAVEFORM
+          0, // OSC2_VOL
+          140, // OSC2_SEMI
+          0, // OSC2_DETUNE
+          0, // OSC2_XENV
+          64, // NOISE_VOL
+          5, // ENV_ATTACK
+          0, // ENV_SUSTAIN
+          67, // ENV_RELEASE
+          104, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          0, // LFO_AMT
+          0, // LFO_FREQ
+          0, // LFO_FX_FREQ
+          3, // FX_FILTER
+          161, // FX_FREQ
+          192, // FX_RESONANCE
+          0, // FX_DIST
+          32, // FX_DRIVE
+          0, // FX_PAN_AMT
+          0, // FX_PAN_FREQ
+          52, // FX_DELAY_AMT
+          1 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,1,1,1],
+          // Columns
+          c: [
+            {n: [147,,147,147,147,,147,,147,,147,147,147,,147,,147,,147,147,147,,147,,147,,147,147,147,,147],
+             f: []}
+          ]
+        },
+        { // Instrument 3
+          i: [
+          3, // OSC1_WAVEFORM
+          73, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          2, // OSC2_WAVEFORM
+          75, // OSC2_VOL
+          128, // OSC2_SEMI
+          6, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          12, // ENV_ATTACK
+          12, // ENV_SUSTAIN
+          33, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          61, // LFO_AMT
+          4, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          109, // FX_FREQ
+          86, // FX_RESONANCE
+          7, // FX_DIST
+          32, // FX_DRIVE
+          112, // FX_PAN_AMT
+          3, // FX_PAN_FREQ
+          67, // FX_DELAY_AMT
+          2 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,2,3,4],
+          // Columns
+          c: [
+            {n: [124,,124,,,,,,,,121,,,,,,120,,132,,,,,,,,114],
+             f: []},
+            {n: [115,,122,,,,,,,,123,,,,,,122,,129,,,,,,,,129],
+             f: []},
+            {n: [124,,131,,,,,,,,136,,,,,,120,,120,,,,,,,,127],
+             f: []},
+            {n: [115,,115,,,,,,,,123,,,,,,122,,122,,,,,,,,123],
+             f: []}
+          ]
+        },
+        { // Instrument 4
+          i: [
+          3, // OSC1_WAVEFORM
+          27, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          2, // OSC2_WAVEFORM
+          27, // OSC2_VOL
+          128, // OSC2_SEMI
+          6, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          12, // ENV_ATTACK
+          12, // ENV_SUSTAIN
+          33, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          61, // LFO_AMT
+          4, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          109, // FX_FREQ
+          86, // FX_RESONANCE
+          7, // FX_DIST
+          32, // FX_DRIVE
+          112, // FX_PAN_AMT
+          3, // FX_PAN_FREQ
+          67, // FX_DELAY_AMT
+          2 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,2,3,4],
+          // Columns
+          c: [
+            {n: [,,,,,,,,,,,,,,,,,,,,,,148,,,,,,,,148,,,,,,,,,,,,,,,,,,,,,,,,151,,,,,,,,151,,,,,,,,,,,,,,,,,,,,,,,,156,,,,,,,,156],
+             f: []},
+            {n: [,,,,,,146,,,,,,,,146,,,,,,,,,,,,,,,,,,,,,,,,151,,,,,,,,151,,,,,,,,,,,,,,,,,,,,,,,,155,,,,,,,,155],
+             f: []},
+            {n: [,,,,,,148,,,,,,,,148,,,,,,,,,,,,,,,,,,,,,,,,151,,,,,,,,151,,,,,,,,,,,,,,,,,,,,,,,,155,,,,,,,,155],
+             f: []},
+            {n: [,,,,,,146,,,,,,,,146,,,,,,,,146,,,,,,,,,,,,,,,,151,,,,,,,,151,,,,,,,,150,,,,,,,,,,,,,,,,155,,,,,,,,155,,,,,,,,153],
+             f: []}
+          ]
+        },
+        { // Instrument 5
+          i: [
+          1, // OSC1_WAVEFORM
+          57, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          1, // OSC2_WAVEFORM
+          57, // OSC2_VOL
+          128, // OSC2_SEMI
+          9, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          5, // ENV_ATTACK
+          28, // ENV_SUSTAIN
+          30, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          195, // LFO_AMT
+          6, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          135, // FX_FREQ
+          0, // FX_RESONANCE
+          0, // FX_DIST
+          18, // FX_DRIVE
+          147, // FX_PAN_AMT
+          6, // FX_PAN_FREQ
+          49, // FX_DELAY_AMT
+          3 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,2,3,4],
+          // Columns
+          c: [
+            {n: [155,,,,,,151,,148,,148,,,,,,156,,,,150,,151,,144],
+             f: []},
+            {n: [158,,,,153,,155,,151,,,,,,,,150,,,,150,,151],
+             f: []},
+            {n: [155,,,,150,,151,,148,,148,,,,,,156,,,,150,,151,,144,,148],
+             f: []},
+            {n: [155,,,,150,,151,,146,,148,,150,,,,150,,,,150,,151],
+             f: []}
+          ]
+        },
+      ],
+      rowLen: 4531,   // In sample lengths
+      patternLen: 32,  // Rows per pattern
+      endPattern: 3,  // End pattern
+      numChannels: 6  // Number of channels
+};