+++ tools/exports/song-title.js
@@ -1,0 +1,143 @@
+// This music has been exported by Voxby, a tracker built on SoundBox.
+// You can use it with https://sb.bitsnbites.eu/player-small.js in your
+// own product; see https://sb.bitsnbites.eu/demo.html for an example.
+
+// Menu · 100 bpm · 8 bars · seed 1589023739
+
+// Song data
+export default {
+      songData: [
+        { // Instrument 0
+          i: [
+          2, // OSC1_WAVEFORM
+          234, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          3, // OSC2_WAVEFORM
+          255, // OSC2_VOL
+          128, // OSC2_SEMI
+          0, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          0, // ENV_ATTACK
+          6, // ENV_SUSTAIN
+          29, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          195, // LFO_AMT
+          4, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          3, // FX_FILTER
+          50, // FX_FREQ
+          184, // FX_RESONANCE
+          119, // FX_DIST
+          244, // FX_DRIVE
+          147, // FX_PAN_AMT
+          6, // FX_PAN_FREQ
+          84, // FX_DELAY_AMT
+          6 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,1,1,1],
+          // Columns
+          c: [
+            {n: [124,,,,,,,,,,,,,,,,124],
+             f: []}
+          ]
+        },
+        { // Instrument 1
+          i: [
+          3, // OSC1_WAVEFORM
+          33, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          2, // OSC2_WAVEFORM
+          33, // OSC2_VOL
+          128, // OSC2_SEMI
+          8, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          14, // ENV_ATTACK
+          34, // ENV_SUSTAIN
+          78, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          61, // LFO_AMT
+          4, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          116, // FX_FREQ
+          72, // FX_RESONANCE
+          4, // FX_DIST
+          32, // FX_DRIVE
+          112, // FX_PAN_AMT
+          3, // FX_PAN_FREQ
+          74, // FX_DELAY_AMT
+          3 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,2,3,2],
+          // Columns
+          c: [
+            {n: [136,,,,136,,,,136,,,,136,,,,136,,,,136,,,,136,,,,136,,,,143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162],
+             f: []},
+            {n: [146,,,,146,,,,146,,,,146,,,,146,,,,146,,,,146,,,,146,,,,153,,,,153,,,,153,,,,153,,,,153,,,,153,,,,153,,,,153,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,172,,,,172,,,,172,,,,172,,,,172,,,,172,,,,172,,,,172],
+             f: []},
+            {n: [143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,143,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,152,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,162,,,,148,,,,148,,,,148,,,,148,,,,148,,,,148,,,,148,,,,148],
+             f: []}
+          ]
+        },
+        { // Instrument 2
+          i: [
+          1, // OSC1_WAVEFORM
+          121, // OSC1_VOL
+          128, // OSC1_SEMI
+          0, // OSC1_XENV
+          3, // OSC2_WAVEFORM
+          150, // OSC2_VOL
+          128, // OSC2_SEMI
+          0, // OSC2_DETUNE
+          0, // OSC2_XENV
+          0, // NOISE_VOL
+          5, // ENV_ATTACK
+          6, // ENV_SUSTAIN
+          58, // ENV_RELEASE
+          0, // ENV_EXP_DECAY
+          0, // ARP_CHORD
+          0, // ARP_SPEED
+          0, // LFO_WAVEFORM
+          195, // LFO_AMT
+          6, // LFO_FREQ
+          1, // LFO_FX_FREQ
+          2, // FX_FILTER
+          135, // FX_FREQ
+          0, // FX_RESONANCE
+          0, // FX_DIST
+          32, // FX_DRIVE
+          147, // FX_PAN_AMT
+          6, // FX_PAN_FREQ
+          121, // FX_DELAY_AMT
+          6 // FX_DELAY_TIME
+          ],
+          // Patterns
+          p: [1,,2,3],
+          // Columns
+          c: [
+            {n: [150,,,,,,,,,,,,153],
+             f: []},
+            {n: [,,,,,,,,,,152,,,,,,,,150,,,,,,,,,,152,,153],
+             f: []},
+            {n: [,,,,,,,,,,,,153,,,,148],
+             f: []}
+          ]
+        },
+      ],
+      rowLen: 6615,   // In sample lengths
+      patternLen: 32,  // Rows per pattern
+      endPattern: 3,  // End pattern
+      numChannels: 3  // Number of channels
+};